Remove example code from readme
The code is already present in the "SOS" example sketch. Having duplicated content is a maintenance liability because it
is likely that one copy or the other will be forgotten when making future changes, resulting in the two falling out of
sync with each other and missing out on enhancements or corrections.